Sort pages before processing
If a page has the 'sval' attribute set, it is used to sort the list of
pages before replacing inlined Python code in page source files. The
default value for 'sval' is 0.

The sorting is useful when a page should get processed before or after
antother one, e.g. when inline Python code updates the attributes of a
page one usually reuses them in another page and that page should get
processed later.