Support custom converters in the macros module
It is now possible to define custom converters in the macros module
(e.g. for SCSS or LESS files). Converters must be defined using a
dictionary named 'converter' with file name patterns as keys and a
converter funcion as well as a target file name extension as values.
Example:

    converter = {
        r'\.less': (less_to_css, 'css'),
        ...
    }

The converter function `less_to_css` must accept the source file name
and the desitnation file name as arguments. The destination file name is
a suggestion, it is the source filename mapped to the project's output
directory with the extension given in the converter dictionary. Example:

    def less_to_css(src, dst):
        # when `src` is '/path/to/project/input/foo.less' then `dst` is
        # '/path/to/project/output/foo.css'
        ...