Python 2/3 hack for the exec statement/function
See http://docs.python.org/release/3.0.1/whatsnew/3.0.html for changes
with regard to `exec`. The problem here is that we cannot use the
function syntax in Poole already for Python 2.x since we need a
qualified exec, and thus we need the `in` syntax which causes a syntax
error in Py3.

This workaround uses `execfile` in Py2 (which, as a function, does not
cause syntax errors in Py3) and the `exec` function from the builtins in
Py3.

References issue 8.